requirements.txt
Django==5.2.1
djangorestframework
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers==3.0.0
transformers
torch
celery[redis]
redis
pgvector
psycopg2-binary
numpy
pandas
tqdm
django-environ
requests
coreschema
django-crispy-forms
django-crispy-forms
pymarc
qdrant-client>=1.8.0